fix(calendar): share recurrence-matching logic between server and ViewMeeting popup
ViewMeeting.tsx's doesMeetingOccurOnDate was a second, independent
implementation of the recurrence predicate meetingOccurrences.ts already
has, and it disagreed with it: it returned true unconditionally for any
monthly pattern, ignored excludedDates and endDate entirely, and
anchored weekly interval differently from the server's Sunday
week-start alignment. This let a recurring meeting's detail popup
re-anchor onto a date the calendar didn't actually render as an
occurrence.

Extract the pure, DB-free predicate logic (matchesRecurrencePattern,
isDateSuspended, addOneETDay, adjustOccurrenceToDate) into
util/meetings/recurrenceMatch.ts, which has no server-only or Prisma
import so client components can use it directly.
meetingOccurrences.ts re-exports from it unchanged for existing
callers, and ViewMeeting.tsx now delegates to the same
matchesRecurrencePattern instead of its own copy.
